Update Swarm Shared Directory Path
Update the shared directory path to be correct and only execute shutil.rmtree if the directory exists to prevent an exception.

R=maruel@chromium.org
BUG=
TEST=Swarm slaves save their output onto the shared drive


Review URL: https://chromiumcodereview.appspot.com/10381126

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@136876 0039d316-1c4b-4281-b951-d872f2087c98

+++ scripts/slave/run_slavelastic.py
@@ -208,12 +208,17 @@ def main():
   if not options.data_dest_dir:
     parser.error('Must specify the server directory')
 
-  # Remove the old data
-  print 'Removing old swarm files...'
-  shutil.rmtree(options.data_dest_dir)
+  # Remove the old data if there is any
+  if os.path.isdir(options.data_dest_dir):
+    print 'Removing old swarm files...'
+    shutil.rmtree(options.data_dest_dir)
 
   # Copy over the new data
   print 'Moving hashtable files to server...'
+  if not os.path.isdir(options.hashtable_dir):
+    print ('ERROR: Given hashtable directory, %s, does not exist' %
+           options.hashtable_dir)
+    return 1
   shutil.copytree(options.hashtable_dir, options.data_path)
 
   # Send off the swarm test requests.
